feat: dissallow deleting itself

+++ app/Containers/AppSection/User/Actions/DeleteUserAction.php
@@ -3,6 +3,7 @@
 namespace App\Containers\AppSection\User\Actions;
 
 use App\Containers\AppSection\User\Data\Repositories\UserRepository;
+use App\Containers\AppSection\User\Exceptions\FailedToDeleteUser;
 use App\Containers\AppSection\User\UI\API\Requests\DeleteUserRequest;
 use App\Ship\Exceptions\ResourceNotFound;
 use App\Ship\Parents\Actions\Action as ParentAction;
@@ -15,10 +16,17 @@ class DeleteUserAction extends ParentAction
     }
 
     /**
+     * @throws FailedToDeleteUser
      * @throws ResourceNotFound
      */
     public function run(DeleteUserRequest $request): bool
     {
+        $user = $this->repository->findById($request->user_id);
+
+        if ($request->user()->is($user)) {
+            throw FailedToDeleteUser::becauseCannotDeleteItself();
+        }
+
         return $this->repository->delete($request->user_id);
     }
 }
